# winx-api/winx-app/app/Models/Employee.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;

class Employee extends Model
{
    /**
     * The attributes that are mass assignable.
     *
     * @var array<int, string>
     */
    protected $fillable = [
        'responsibility',
        'admission_at',
        'phone',
        'user_id',
    ];

    /**
     * The attributes that should be mutated to dates.
     *
     * @var array
     */
    protected $dates = ['admission_at'];
    
    /**
     * Get the User.
     */
    public function user()
    {
        return $this->belongsTo(User::class);
    }
}